requirements.txt
Flask
mysql-connector
pyyaml
